+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Plutus/Context.hs
@@ -13,6 +13,13 @@ module Cardano.Ledger.Alonzo.Plutus.Context (
   EraPlutusTxInfo (..),
   EraPlutusContext (..),
   mkPlutusLanguageContext,
+
+  -- * Language dependent translation
+  PlutusTxInfo,
+  PlutusTxCert,
+  PlutusTxOut,
+  PlutusScriptPurpose,
+  PlutusScriptContext,
 )
 where
 
@@ -22,7 +29,6 @@ import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Core
 import Cardano.Ledger.Plutus.Data (Data (..))
 import Cardano.Ledger.Plutus.Language (Language (..), PlutusLanguage)
-import Cardano.Ledger.Plutus.TxInfo
 import Cardano.Ledger.UTxO (UTxO (..))
 import Cardano.Slotting.EpochInfo (EpochInfo)
 import Cardano.Slotting.Time (SystemStart)
@@ -31,6 +37,9 @@ import Data.Kind (Type)
 import Data.Text (Text)
 import NoThunks.Class (NoThunks)
 import qualified PlutusLedgerApi.V1 as P (ToData, toData)
+import qualified PlutusLedgerApi.V1 as PV1
+import qualified PlutusLedgerApi.V2 as PV2
+import qualified PlutusLedgerApi.V3 as PV3
 
 class (PlutusLanguage l, AlonzoEraScript era) => EraPlutusTxInfo (l :: Language) era where
   toPlutusTxCert :: proxy l -> TxCert era -> Either (ContextError era) (PlutusTxCert l)
@@ -92,3 +101,31 @@ mkPlutusLanguageContext proxy scriptPurpose pp epochInfo sysStart utxo tx = do
   txInfo <- toPlutusTxInfo proxy pp epochInfo sysStart utxo tx
   scriptContext <- toPlutusScriptContext proxy txInfo scriptPurpose
   pure $ Data $ P.toData scriptContext
+
+-- =============================================
+-- Type families that specify Plutus types that are different from one version to another
+
+type family PlutusTxCert (l :: Language) where
+  PlutusTxCert 'PlutusV1 = PV1.DCert
+  PlutusTxCert 'PlutusV2 = PV2.DCert
+  PlutusTxCert 'PlutusV3 = PV3.TxCert
+
+type family PlutusTxOut (l :: Language) where
+  PlutusTxOut 'PlutusV1 = PV1.TxOut
+  PlutusTxOut 'PlutusV2 = PV2.TxOut
+  PlutusTxOut 'PlutusV3 = PV3.TxOut
+
+type family PlutusScriptPurpose (l :: Language) where
+  PlutusScriptPurpose 'PlutusV1 = PV1.ScriptPurpose
+  PlutusScriptPurpose 'PlutusV2 = PV2.ScriptPurpose
+  PlutusScriptPurpose 'PlutusV3 = PV3.ScriptPurpose
+
+type family PlutusScriptContext (l :: Language) where
+  PlutusScriptContext 'PlutusV1 = PV1.ScriptContext
+  PlutusScriptContext 'PlutusV2 = PV2.ScriptContext
+  PlutusScriptContext 'PlutusV3 = PV3.ScriptContext
+
+type family PlutusTxInfo (l :: Language) where
+  PlutusTxInfo 'PlutusV1 = PV1.TxInfo
+  PlutusTxInfo 'PlutusV2 = PV2.TxInfo
+  PlutusTxInfo 'PlutusV3 = PV3.TxInfo
