+++ eras/babbage/test-suite/src/Test/Cardano/Ledger/Babbage/TxInfo.hs
@@ -10,7 +10,7 @@ module Test.Cardano.Ledger.Babbage.TxInfo where
 
 import Cardano.Ledger.Address (Addr (..), BootstrapAddress (..))
 import Cardano.Ledger.Alonzo.PParams ()
-import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo (toPlutusTxInfo))
+import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo (toPlutusTxInfo), PlutusTxInfo)
 import Cardano.Ledger.Alonzo.Plutus.TxInfo (TxOutSource (..))
 import Cardano.Ledger.Alonzo.Scripts (AlonzoEraScript, AlonzoScript)
 import Cardano.Ledger.Babbage (Babbage)
@@ -26,7 +26,6 @@ import Cardano.Ledger.Crypto (Crypto)
 import Cardano.Ledger.Mary.Value (MaryValue)
 import Cardano.Ledger.Plutus.Data (Data (..), dataToBinaryData)
 import Cardano.Ledger.Plutus.Language (Language (..), SLanguage (..))
-import Cardano.Ledger.Plutus.TxInfo (PlutusTxInfo)
 import Cardano.Ledger.TxIn (TxIn (..), mkTxInPartial)
 import Cardano.Ledger.UTxO (UTxO (..))
 import qualified Cardano.Ledger.Val as Val
